refactor: Move Section components into their own files
