+++ src/components/EditionModal/index.jsx
@@ -1,0 +1,3 @@
+export { default as AccountOrGroupSection } from './AccountOrGroupSection'
+export { default as CategorySection } from './CategorySection'
+export { default as ThresholdSection } from './ThresholdSection'
